refactor: Move file location for readability
가독성을 위해 파일 구조 변경
